refactor : edit api module overall because of board extends #12
