--- Trim-Api/src/main/java/trim/api/ApiPackageLocation.java
@@ -1,4 +1,0 @@
-package trim.api;
-
-public interface ApiPackageLocation {
-}
