+++ Trim-Api/src/main/java/trim/api/common/advice/ExceptionAdvice.java
@@ -14,12 +14,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.ServletWebRequest;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import trim.api.ApiPackageLocation;
 import trim.api.common.dto.ApiResponseDto;
-import trim.api.common.example.ExampleApiController;
-import trim.api.domains.comment.controller.QuestionCommentApiController;
-import trim.api.domains.member.controller.MemberApiController;
-import trim.api.domains.question.controller.QuestionApiController;
 import trim.common.exception.ErrorStatus;
 import trim.common.exception.GeneralException;
 import trim.common.exception.Reason;
