+++ Trim-Api/src/main/java/trim/api/domains/comment/controller/QuestionCommentApiController.java
@@ -5,7 +5,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import trim.api.common.dto.ApiResponseDto;
 import trim.api.domains.comment.service.CreateQuestionCommentUseCase;
-import trim.api.domains.comment.service.DeleteQuestionCommentUseCase;
 import trim.api.domains.comment.vo.request.CreateQuestionCommentRequest;
 import trim.domains.member.dao.domain.Member;
 
@@ -15,8 +14,7 @@ import trim.domains.member.dao.domain.Member;
 public class QuestionCommentApiController {
 
     private final CreateQuestionCommentUseCase createQuestionCommentUseCaseUseCase;
-    private final DeleteQuestionCommentUseCase deleteQuestionCommentUseCase;
-    
+
     /** 댓글 등록 테스트용 **/
     @PostMapping
     public ApiResponseDto<Long> createQuestionComment(Member writer,
@@ -24,14 +22,6 @@ public class QuestionCommentApiController {
                                                 @RequestBody CreateQuestionCommentRequest dto){
 
         return ApiResponseDto.onSuccess(createQuestionCommentUseCaseUseCase.execute(writer, questionId, dto));
-    }
-
-    /** 댓글 삭제 테스트용 **/
-    @DeleteMapping("/{commentId}")
-    public ApiResponseDto<Boolean> deleteQuestionComment(Member writer,
-                                      @PathVariable("commentId") Long commentId){
-        deleteQuestionCommentUseCase.execute(writer, commentId);
-        return ApiResponseDto.onSuccess(true);
     }
 
 }
