+++ Trim-Api/src/main/java/trim/api/domains/comment/service/CreateQuestionCommentUseCase.java
@@ -3,17 +3,18 @@ package trim.api.domains.comment.service;
 import lombok.RequiredArgsConstructor;
 import trim.api.domains.comment.vo.request.CreateQuestionCommentRequest;
 import trim.common.annotation.UseCase;
-import trim.domains.comment.business.service.QuestionCommentDomainService;
+import trim.domains.board.business.adaptor.QuestionAdaptor;
+import trim.domains.board.dao.domain.Question;
+import trim.domains.comment.business.service.CommentDomainService;
 import trim.domains.member.dao.domain.Member;
-import trim.domains.question.business.adaptor.QuestionAdaptor;
-import trim.domains.question.dao.domain.Question;
+
 
 @UseCase
 @RequiredArgsConstructor
 public class CreateQuestionCommentUseCase {
 
     private final QuestionAdaptor questionAdaptor;
-    private final QuestionCommentDomainService questionCommentDomainService;
+    private final CommentDomainService questionCommentDomainService;
 
     public Long execute(Member writer, Long questionId, CreateQuestionCommentRequest dto){
         Question question = questionAdaptor.queryById(questionId);
