+++ Trim-Api/src/main/java/trim/api/domains/comment/vo/response/QuestionCommentResponse.java
@@ -2,7 +2,7 @@ package trim.api.domains.comment.vo.response;
 
 import lombok.Builder;
 import lombok.Getter;
-import trim.domains.comment.dao.domain.QuestionComment;
+import trim.domains.comment.dao.domain.Comment;
 
 @Getter
 @Builder
@@ -10,10 +10,10 @@ public class QuestionCommentResponse {
     private String content;
     private String nickname;
 
-    public static QuestionCommentResponse of(QuestionComment questionComment){
+    public static QuestionCommentResponse of(Comment comment){
        return QuestionCommentResponse.builder()
-               .content(questionComment.getContent())
-               .nickname(questionComment.getWriter().getNickname())
+               .content(comment.getContent())
+               .nickname(comment.getWriter().getNickname())
                .build();
     }
 }
