+++ Trim-Api/src/main/java/trim/api/domains/question/dto/response/QuestionResponse.java
@@ -3,8 +3,8 @@ package trim.api.domains.question.dto.response;
 import lombok.Builder;
 import lombok.Getter;
 import trim.api.domains.comment.vo.response.QuestionCommentResponse;
-import trim.domains.comment.dao.domain.QuestionComment;
-import trim.domains.question.dao.domain.Question;
+import trim.domains.board.dao.domain.Question;
+import trim.domains.comment.dao.domain.Comment;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -19,7 +19,7 @@ public class QuestionResponse {
 
     private List<QuestionCommentResponse> questionComments;
 
-    public static QuestionResponse of(Question question, List<QuestionComment> questionComments) {
+    public static QuestionResponse of(Question question, List<Comment> questionComments) {
         return QuestionResponse.builder()
                 .title(question.getTitle())
                 .content(question.getContent())
