+++ Trim-Api/src/main/java/trim/api/domains/question/service/EditQuestionUseCase.java
@@ -3,11 +3,11 @@ package trim.api.domains.question.service;
 import lombok.RequiredArgsConstructor;
 import trim.api.domains.question.dto.request.QuestionRequest;
 import trim.common.annotation.UseCase;
+import trim.domains.board.business.adaptor.QuestionAdaptor;
+import trim.domains.board.business.service.QuestionDomainService;
+import trim.domains.board.business.validate.QuestionValidate;
+import trim.domains.board.dao.domain.Question;
 import trim.domains.member.dao.domain.Member;
-import trim.domains.question.business.adaptor.QuestionAdaptor;
-import trim.domains.question.dao.domain.Question;
-import trim.domains.question.business.service.QuestionDomainService;
-import trim.domains.question.business.validate.QuestionValidate;
 
 @UseCase
 @RequiredArgsConstructor
