+++ Trim-Api/src/main/java/trim/api/domains/question/service/FindQuestionUseCase.java
@@ -3,10 +3,10 @@ package trim.api.domains.question.service;
 import lombok.RequiredArgsConstructor;
 import trim.api.domains.question.dto.response.QuestionResponse;
 import trim.common.annotation.UseCase;
-import trim.domains.comment.business.adaptor.QuestionCommentAdaptor;
-import trim.domains.comment.dao.domain.QuestionComment;
-import trim.domains.question.business.adaptor.QuestionAdaptor;
-import trim.domains.question.dao.domain.Question;
+import trim.domains.board.business.adaptor.QuestionAdaptor;
+import trim.domains.board.dao.domain.Question;
+import trim.domains.comment.business.adaptor.CommentAdaptor;
+import trim.domains.comment.dao.domain.Comment;
 
 import java.util.List;
 
@@ -15,11 +15,11 @@ import java.util.List;
 public class FindQuestionUseCase {
 
     private final QuestionAdaptor questionAdaptor;
-    private final QuestionCommentAdaptor questionCommentAdaptor;
+    private final CommentAdaptor commentAdaptor;
 
     public QuestionResponse execute(Long questionId) {
         Question question = questionAdaptor.queryById(questionId);
-        List<QuestionComment> questionComments = questionCommentAdaptor.queryAllByQuestionId(questionId);
-        return QuestionResponse.of(question, questionComments);
+        List<Comment> comments = commentAdaptor.queryAllByBoardId(questionId);
+        return QuestionResponse.of(question, comments);
     }
 }
