+++ Trim-Domain/src/main/java/trim/domains/board/business/validate/QuestionValidate.java
@@ -1,7 +1,7 @@
-package trim.domains.question.business.validate;
+package trim.domains.board.business.validate;
 
+import trim.domains.board.dao.domain.Question;
 import trim.domains.member.dao.domain.Member;
-import trim.domains.question.dao.domain.Question;
 
 public interface QuestionValidate {
